requirements.txt
ffmpeg-python==0.2.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.40.1rc1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper==20230314
regex==2023.6.3
requests==2.31.0
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
urllib3==2.0.3
Werkzeug==2.3.6